% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_engine.R
\name{.cleanup_checkpoint}
\alias{.cleanup_checkpoint}
\title{Cleanup Checkpoint File}
\usage{
.cleanup_checkpoint(session_id)
}
\arguments{
\item{session_id}{Character session identifier.}
}
\description{
Cleanup Checkpoint File
}
\keyword{internal}
