% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_map_all.R
\name{.create_safe_future_map}
\alias{.create_safe_future_map}
\title{Create Safe Future Map Variant}
\usage{
.create_safe_future_map(
  mode,
  output_type = "list",
  transform_fn = NULL,
  bind_fn = NULL
)
}
\description{
Create Safe Future Map Variant
}
\keyword{internal}
