% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_map_all.R
\name{.create_safe_map2}
\alias{.create_safe_map2}
\title{Create Safe Map2 Function Variant}
\usage{
.create_safe_map2(
  mode,
  output_type = "list",
  transform_fn = NULL,
  bind_fn = NULL
)
}
\description{
Create Safe Map2 Function Variant
}
\keyword{internal}
