% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_engine.R
\name{.execute_batch}
\alias{.execute_batch}
\title{Execute Single Batch}
\usage{
.execute_batch(
  data,
  func,
  batch_indices,
  mode,
  .options = NULL,
  .env_globals = NULL,
  .progress = FALSE,
  ...
)
}
\arguments{
\item{data}{List of input data.}

\item{func}{Function to apply.}

\item{batch_indices}{Integer vector of indices for this batch.}

\item{mode}{Character operation mode.}

\item{...}{Additional arguments passed to func.}
}
\value{
List of batch results.
}
\description{
Execute Single Batch
}
\keyword{internal}
