% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_engine.R
\name{.format_output}
\alias{.format_output}
\title{Format Output According to Type}
\usage{
.format_output(results, output_type)
}
\arguments{
\item{results}{List of results.}

\item{output_type}{Character output type.}
}
\value{
Formatted output.
}
\description{
Format Output According to Type
}
\keyword{internal}
