% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_engine.R
\name{.get_checkpoint_path}
\alias{.get_checkpoint_path}
\title{Get Checkpoint File Path}
\usage{
.get_checkpoint_path(session_id)
}
\arguments{
\item{session_id}{Character session identifier.}
}
\value{
Character file path.
}
\description{
Get Checkpoint File Path
}
\keyword{internal}
