% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_engine.R
\name{.save_checkpoint}
\alias{.save_checkpoint}
\title{Save Checkpoint}
\usage{
.save_checkpoint(session_id, results, total_items, mode, completed_idx)
}
\arguments{
\item{session_id}{Character session identifier.}

\item{results}{List of results so far.}

\item{total_items}{Integer total number of items.}

\item{mode}{Character operation mode.}

\item{completed_idx}{Integer index of last completed item.}
}
\description{
Save Checkpoint
}
\keyword{internal}
