% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_engine.R
\name{null-default}
\alias{null-default}
\alias{grapes-or-or-grapes}
\alias{\%||\%}
\title{Null-coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value}

\item{y}{Alternative value if x is NULL}
}
\value{
x if not NULL, otherwise y
}
\description{
Null-coalescing Operator
}
\keyword{internal}
