% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_map_all.R
\name{s_future_map2}
\alias{s_future_map2}
\alias{s_future_map2_chr}
\alias{s_future_map2_dbl}
\alias{s_future_map2_int}
\alias{s_future_map2_lgl}
\alias{s_future_map2_dfr}
\alias{s_future_map2_dfc}
\title{Safe Future Map2 - Parallel Two-Input with Auto-Recovery}
\usage{
s_future_map2(
  .x,
  .y,
  .f,
  ...,
  .options = NULL,
  .env_globals = parent.frame(),
  .progress = FALSE,
  .id = NULL,
  .session_id = NULL
)

s_future_map2_chr(
  .x,
  .y,
  .f,
  ...,
  .options = NULL,
  .env_globals = parent.frame(),
  .progress = FALSE,
  .id = NULL,
  .session_id = NULL
)

s_future_map2_dbl(
  .x,
  .y,
  .f,
  ...,
  .options = NULL,
  .env_globals = parent.frame(),
  .progress = FALSE,
  .id = NULL,
  .session_id = NULL
)

s_future_map2_int(
  .x,
  .y,
  .f,
  ...,
  .options = NULL,
  .env_globals = parent.frame(),
  .progress = FALSE,
  .id = NULL,
  .session_id = NULL
)

s_future_map2_lgl(
  .x,
  .y,
  .f,
  ...,
  .options = NULL,
  .env_globals = parent.frame(),
  .progress = FALSE,
  .id = NULL,
  .session_id = NULL
)

s_future_map2_dfr(
  .x,
  .y,
  .f,
  ...,
  .options = NULL,
  .env_globals = parent.frame(),
  .progress = FALSE,
  .id = NULL,
  .session_id = NULL
)

s_future_map2_dfc(
  .x,
  .y,
  .f,
  ...,
  .options = NULL,
  .env_globals = parent.frame(),
  .progress = FALSE,
  .id = NULL,
  .session_id = NULL
)
}
\arguments{
\item{.x, .y}{Vectors of the same length.}

\item{.f}{A function, formula, or vector.}

\item{...}{Additional arguments passed to .f.}

\item{.options}{A furrr_options object.}

\item{.env_globals}{The environment to look for globals.}

\item{.progress}{A single logical.}

\item{.id}{Optional name for ID column (dfr/dfc variants).}

\item{.session_id}{Character. Optional session ID.}
}
\value{
A list.
}
\description{
Safe Future Map2 - Parallel Two-Input with Auto-Recovery
}
