% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_other.R
\name{s_imap}
\alias{s_imap}
\alias{s_imap_chr}
\alias{s_future_imap}
\title{Safe IMap - Drop-in Replacement for purrr::imap with Auto-Recovery}
\usage{
s_imap(.x, .f, ..., .session_id = NULL)

s_imap_chr(.x, .f, ..., .session_id = NULL)

s_future_imap(
  .x,
  .f,
  ...,
  .options = NULL,
  .env_globals = parent.frame(),
  .progress = FALSE,
  .session_id = NULL
)
}
\arguments{
\item{.x}{A list or atomic vector.}

\item{.f}{A function, formula, or vector.}

\item{...}{Additional arguments passed to .f.}

\item{.session_id}{Character. Optional session ID.}

\item{.options}{A furrr_options object (NULL uses defaults).}

\item{.env_globals}{The environment to look for globals.}

\item{.progress}{A single logical.}
}
\value{
A list.

A character vector.
}
\description{
Safe IMap - Drop-in Replacement for purrr::imap with Auto-Recovery
}
