% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugging.R
\name{s_session_stats}
\alias{s_session_stats}
\title{Get SafeMapper Session Statistics}
\usage{
s_session_stats()
}
\value{
Invisible data frame of session information.
}
\description{
Provides summary statistics of all checkpoint sessions.
}
\keyword{internal}
