% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugging.R
\name{s_version}
\alias{s_version}
\title{Get SafeMapper Package Version and Information}
\usage{
s_version()
}
\description{
Displays comprehensive information about the SafeMapper package including
version, supported functions, and usage tips.
}
\examples{
# Show package information
s_version()

}
