% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_other.R
\name{s_pmap}
\alias{s_pmap}
\alias{s_future_pmap}
\title{Safe PMap - Drop-in Replacement for purrr::pmap with Auto-Recovery}
\usage{
s_pmap(.l, .f, ..., .session_id = NULL)

s_future_pmap(
  .l,
  .f,
  ...,
  .options = furrr::furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE,
  .session_id = NULL
)
}
\arguments{
\item{.l}{A list of lists/vectors to map over.}

\item{.f}{A function, formula, or vector.}

\item{...}{Additional arguments passed to .f.}

\item{.session_id}{Character. Optional session ID.}

\item{.options}{A furrr_options object (for s_future_pmap only).}

\item{.env_globals}{The environment to look for globals (for s_future_pmap only).}

\item{.progress}{A single logical for progress bar (for s_future_pmap only).}
}
\value{
A list.
}
\description{
Safe PMap - Drop-in Replacement for purrr::pmap with Auto-Recovery
}
